--- backend/controller/leases/dal/lease.go
@@ -1,1 +1,0 @@
-package dal
